3-3

+++ CodingInterviewJava/src/codinginterviewjava/Chapter2.java
@@ -28,7 +28,7 @@ class Node {
     }
 
     Node(Object item) {
-        data_ = (int)item;
+        data_ = item;
     }
 
     Node appendToTail(Node add) {
